+++ voicevox_engine/tts_pipeline/__init__.py
@@ -1,5 +1,5 @@
 from ..core_wrapper import CoreWrapper, load_runtime_lib
-from .make_synthesis_engines import make_synthesis_engines
+from .make_tts_engines import make_synthesis_engines
 from .tts_engine import SynthesisEngine
 from .tts_engine_base import SynthesisEngineBase
